refactor: svgicons 파일명 변경
