+++ frontend/src/components/@common/SvgIcons/SvgSpriteMap.tsx
